requirements.txt
python-telegram-bot==20.7
python-dotenv==1.0.0
langchain==0.1.0
openai==1.3.0
SQLAlchemy==2.0.23
python-dateutil==2.8.2